Make agent capable of multiple intercepts
Divides the agent state in two. One that applies to the agent as such
and then a slice where each entry applies to an intercept. In this
implementation, a single intercept state is created from the environment
so that everything just continues to work as it did before. The intent
is to later on be able to create several such states based on a
`ConfigMap` and hence make the agent support multiple simultaneous
intercepts that uses a different set of ports, environment, and mount
point.

This commit also removes moves the local `mount_point` field from the
gRPC `InterceptSpec` to the `InterceptInfo` and renames it to
`client_mount_point` (naming is consistent with "client_session"). The
reason for this move is that the field isn't part of the static
intercept and clients may actually mount this at different locations.

Signed-off-by: Thomas Hallgren <thomas@tada.se>